add personall info routes authentication

+++ routers/user/personalInfo.ts
@@ -1,108 +1,38 @@
 import {Router} from "express";
-import {PersonalInfoRecord, UserRecord} from "../../records";
-import {PersonalInfoCreateReq, PersonalInfoEntity} from "../../types";
-import {exists, isNotNull, isNull, isSmaller, isTypeOf} from "../../utils/dataCheck";
+import {PersonalInfoRecord} from "../../records";
+import { UserAuthReq} from "../../types";
+import { isNull, isSmaller, isTypeOf} from "../../utils/dataCheck";
 import {SetPersonalInfoReq} from "../../types";
+import {authenticateToken} from "../../middleware/auth";
+import {AuthInvalidError, InvalidTokenError} from "../../utils/errors";
 export const personalInfoRouter = Router();
 
 personalInfoRouter
 
-    .get('/all', async (req, res) => {
-        const personalInfoList = await PersonalInfoRecord.listAll();
+    .get('/one', authenticateToken,async (req: UserAuthReq, res) => {
 
-        res.json({
-            personalInfoList,
-        })
-    })
-    .get('/one/:id', async (req, res) => {
-        exists(req.params.id, 'id param')
-        const personalInfo = await PersonalInfoRecord.getOne(req.params.id);
-        isNull(personalInfo, null,'shopItem does not exists')
+        const { id: reqUserId } = req.user
+        if (!reqUserId) throw new InvalidTokenError()
+
+        const personalInfo = await PersonalInfoRecord.getOne(reqUserId);
+        isNull(personalInfo, null,'personalInfo does not exists')
+
+        if (personalInfo.userId !== reqUserId) throw new AuthInvalidError()
 
         res.json({
             personalInfo,
         })
     })
-    .post('/', async (req, res) => {
-        const { body } : {
-            body: PersonalInfoCreateReq
-        } = req
-
-        exists(body.userId, 'user id')
-        isTypeOf(body.userId, 'string', 'user Id')
-
-        const user = await UserRecord.getOne(body.userId);
-        isNull(user, null,'user does not exists')
-
-        const personalInfo = await PersonalInfoRecord.getUserInfo(body.userId)
-        isNotNull(personalInfo, '', 'this user already has personal info')
-
-        if (body.name) {
-            isTypeOf(body.name, 'string', 'name')
-            isSmaller(body.name.length, 45,'', 'name is longer than 45 characters')
-        } else {
-            body.name = null
-        }
-
-        if (body.surname) {
-            isTypeOf(body.surname, 'string', 'surname')
-            isSmaller(body.surname.length, 47, '', 'surname is longer than 47 characters')
-        } else {
-            body.surname = null
-        }
-
-        if (body.city) {
-            isTypeOf(body.city, 'string', 'city')
-            isSmaller(body.city.length, 85, '', 'city name is longer than 85 characters')
-        } else {
-            body.city = null
-        }
-
-        if (body.country) {
-            isTypeOf(body.country, 'string', 'country')
-            isSmaller(body.country.length, 56, '', 'country nam is longer than 56 characters')
-        } else {
-            body.country = null
-        }
-
-        if (body.street) {
-            isTypeOf(body.street, 'string', 'street')
-            isSmaller(body.street.length, 85, '', 'street name is longer than 85 characters')
-        } else {
-            body.street = null
-        }
-
-        if (body.buildingNumber) {
-            isTypeOf(body.buildingNumber, 'string', 'buildingNumber')
-            isSmaller(body.buildingNumber.length, 10, '', 'building number is longer than 10 characters')
-        } else {
-            body.buildingNumber = null
-        }
-
-        if (body.postalCode) {
-            isTypeOf(body.postalCode, 'string', 'postalCode')
-            isSmaller(body.postalCode.length, 6, '', 'postal code number is longer than 6 characters')
-        } else {
-            body.postalCode = null
-        }
-
-        const newPersonalInfo = new PersonalInfoRecord(req.body as PersonalInfoCreateReq);
-        await newPersonalInfo.insert();
-
-
-        res.json(newPersonalInfo as PersonalInfoEntity);
-    })
-
-    .patch('/:personalInfoId', async (req, res) => {
+       .patch('/',authenticateToken, async (req: UserAuthReq, res) => {
         const { body } : {
             body: SetPersonalInfoReq
         } = req
 
-        exists(req.params.personalInfoId, 'personal Info Id param')
-        const personalInfo = await PersonalInfoRecord.getOne(req.params.personalInfoId);
-        isNull(personalInfo, null,'Personal Info does not exists')
+        const { id: reqUserId } = req.user
+        if (!reqUserId) throw new InvalidTokenError()
 
-        console.log(body.name, 'name')
+        const personalInfo = await PersonalInfoRecord.getUserInfo(reqUserId);
+        isNull(personalInfo, null,'Personal Info does not exists')
 
         if (body.name) {
             isTypeOf(body.name, 'string', 'name')
@@ -151,12 +81,12 @@ personalInfoRouter
         res.json(personalInfo)
     })
 
-    .delete('/:personalInfoId', async (req, res) => {
+    .delete('/', authenticateToken,async (req: UserAuthReq, res) => {
+        const { id: reqUserId } = req.user
+        if (!reqUserId) throw new InvalidTokenError()
 
-        exists(req.params.personalInfoId, 'shop item id param')
-        const personalInfo = await PersonalInfoRecord.getOne(req.params.personalInfoId);
-
-        isNull(personalInfo, null,'No shopItem found for this ID.')
+        const personalInfo = await PersonalInfoRecord.getUserInfo(reqUserId);
+        isNull(personalInfo, null,'No personalInfo found for this ID.')
 
         await personalInfo.delete();
         res.json({message: "personal Info deleted successfully."})
